Feat(NO-TICKET): cleanup folder structure
